Update budget algorithm when changing priority Return and change real priority Fix LWP_Reschedule

+++ gc/ogc/lwp.h
@@ -150,14 +150,14 @@ lwp_t LWP_GetSelf(void);
 s32 LWP_GetThreadPriority(lwp_t thethread);
 
 
-/*! \fn s32 LWP_SetThreadPriority(lwp_t thethread,u32 prio)
+/*! \fn s32 LWP_SetThreadPriority(lwp_t thethread,u8 prio)
 \brief Set the priority of the given thread.
 \param[in] thethread handle to the thread context whose priority should be changed. If NULL, the current thread will be taken.
 \param[in] prio new thread priority to set
 
 \return old thread priority
 */
-s32 LWP_SetThreadPriority(lwp_t thethread,u32 prio);
+s32 LWP_SetThreadPriority(lwp_t thethread,u8 prio);
 
 
 /*! \fn void LWP_YieldThread(void)
@@ -168,13 +168,13 @@ s32 LWP_SetThreadPriority(lwp_t thethread,u32 prio);
 void LWP_YieldThread(void);
 
 
-/*! \fn void LWP_Reschedule(u32 prio)
+/*! \fn void LWP_Reschedule(u8 prio)
 \brief Reschedule all threads running at the given priority
 \param[in] prio priority level to reschedule
 
 \return none
 */
-void LWP_Reschedule(u32 prio);
+void LWP_Reschedule(u8 prio);
 
 
 /*! \fn void LWP_ExitThread(void *value_ptr)
